fix: STORED didn't lift as reveal instruction

+++ extensions/riscv/rvr/c/rv64io_callbacks.h
@@ -8,6 +8,6 @@
  * installed by `Rv64IoExtension::register_host_callbacks`. */
 void openvm_hint_storew(uint32_t dest_addr);
 void openvm_hint_buffer(uint32_t dest_addr, uint32_t num_words);
-void openvm_reveal(uint32_t src_val, uint32_t ptr, uint32_t offset);
+void openvm_reveal(uint64_t src_val, uint32_t ptr, uint32_t offset);
 
 #endif /* RV64IO_CALLBACKS_H */
